+++ src/finiteVolume/cellCentred/operators/ddtOperator.cpp
@@ -228,8 +228,8 @@ void DdtOperator<ValueType>::read(const Input& input)
         return;
     }
 
-    NF_ERROR_EXIT(std::format(
-        "Unknown ddt scheme '{}' for field '{}'. Supported schemes are: BDF1, BDF2.",
+    NF_ERROR_EXIT(fmt::format(
+        fmt::runtime("Unknown ddt scheme '{}' for field '{}'. Supported schemes are: BDF1, BDF2."),
         schemeName,
         this->field_.name
     ));