+++ src/finiteVolume/cellCentred/stencil/cellToFaceStencil.cpp
@@ -40,6 +40,8 @@ SegmentedVector<localIdx, localIdx> CellToFaceStencil::computeInternalStencil() 
 
     fill(nFacesPerCell, 0); // reset nFacesPerCell
 
+    // TODO on GPU this might lead to undefined order
+    // find an implementation that guarantees the correct order
     parallelFor(
         serialExec,
         {0, nInternalFaces},